+++ seleniumtid/utils.py
@@ -57,7 +57,7 @@ class Utils(object):
         '''
         for entry in self.driver.get_log(log_type):
             message = entry['message'].rstrip().encode('utf-8')
-            self.logger.debug('{0} - {1}: {2}'.format(log_type.title(), entry['level'], message))
+            self.logger.debug('{0} - {1}: {2}'.format(log_type.capitalize(), entry['level'], message))
 
     def wait_until_element_not_visible(self, locator, timeout=10):
         '''
@@ -157,3 +157,11 @@ class Utils(object):
         except (requests.exceptions.ConnectionError, KeyError):
             record_videos = 'false'
         return True if record_videos == 'true' else False
+
+
+class classproperty (property):
+    '''
+    Subclass property to make classmethod properties possible
+    '''
+    def __get__(self, cls, owner):
+        return self.fget.__get__(None, owner)()